refactor(activity): extract formatActivity to lib module to fix build
Next.js App Router only permits HTTP-verb exports (GET, POST, etc.)
from route files. Exporting 'formatActivity' directly from route.ts
caused the CI build to fail:

  Route "src/app/api/metrics/activity/route.ts" does not match the
  required types of a Next.js Route.
  "formatActivity" is not a valid Route export field.

Fix:
- Move formatActivity, its types (ActivityItem, RawEvent, ActivityType),
  and SUPPORTED_EVENT_TYPES into src/lib/activity-formatter.ts
- Import from the new module in route.ts (no functional change)
- Update test/formatActivity.test.ts to import from the lib module
  (removes now-unnecessary vi.mock calls for supabase/auth/next-auth)
